WSL: Implement namespaces
Because WSL is a persistent VM, and each distribution is just a
namespace in that VM (but using a shared network namespace), shutting
down Rancher Desktop and restarting it does not always reset things
correctly.  Fix things by re-introducing network namespaces that we
previously used in Rancher Desktop 1.

This also re-introduces the collection of network-related executables we
used to manage the network namespaces.

Buildkit requires cgroups v2 to work correctly.  This means we need to
use a shared mount namespace across all of our processes, so using the
systemd-native `JoinsNamespaceOf=` does not work (it only joins network
namespaces plus some select mounts).  Instead, manually modify the
`ExecStart=` commands to use `/usr/local/bin/wsl-exec` to join the
namespace, in the same way we do for `rdctl shell`.

Signed-off-by: Mark Yen <mark.yen@suse.com>