wsl-exec: Ensure a minimum set of PATH
If we override PATH when spawned via Windows (via setting WSLENV), the
default set of entries do not get picked up because it makes no sense on
the Windows side.  This means that basic system utilities end up missing;
this causes issues like runc not being able to find iptables.

When wsl-exec is run, ensure that a hard-coded list (taken from
`/etc/login.defs`) is found.

Signed-off-by: Mark Yen <mark.yen@suse.com>