CI: Only save cache on main
Our caches are rather big (nearly all of the total quota).  Unless the
cache was built from the default branch, it can only be used from the same
branch as the run that pushed it.  This means that we end up pushing lots
of cache that can never be reused, causing the actually useful ones (i.e.
the caches from the default branch) to be evicted due to space.

Fix this by using explicit cache save and restore, and then limiting the
save to only apply on the main branch.

Signed-off-by: Mark Yen <mark.yen@suse.com>